pyproject.toml
[tool.poetry]
name = "tweet-challenge"
version = "0.1.0"
description = ""
authors = ["Kyle Kim <kyle.kim@nativept.kr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
black = "^24.10.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
